requirements.txt
pandas
fastapi
uvicorn
asyncio
gunicorn
python-dotenv
